address_name_yl/requirements.txt
geopy
jieba
shapely
requests
networkx
zhconv
sanic
gspread
oauth2client
aiohttp
ujson
area
aioelasticsearch
elasticsearch
sanic_openapi
aioredis
motor
python-dotenv